+++ Roslyn~/Coimbra.Roslyn/Diagnostics.cs
@@ -69,7 +69,6 @@ namespace Coimbra.Roslyn
                                                                                                      DiagnosticSeverity.Error,
                                                                                                      true);
 
-        // TODO: add CodeFix
         public static readonly DiagnosticDescriptor EventServiceGenericMethodsShouldNotBeUsedDirectly = new("COIMBRA0009",
                                                                                                             "IEventService generic APIs should not be used directly.",
                                                                                                             "Use {0}.{1} instead.",
@@ -126,7 +125,6 @@ namespace Coimbra.Roslyn
                                                                                                DiagnosticSeverity.Error,
                                                                                                true);
 
-        // TODO: add CodeFix
         public static readonly DiagnosticDescriptor OwningLocatorShouldBeUsedInsteadOfShared = new("COIMBRA017",
                                                                                                    "ServiceLocator.Shared should not be accessed inside IService.",
                                                                                                    "Use OwningLocator instead of ServiceLocator.Shared.",
@@ -141,7 +139,6 @@ namespace Coimbra.Roslyn
                                                                                                   DiagnosticSeverity.Error,
                                                                                                   true);
 
-        // TODO: add CodeFix
         public static readonly DiagnosticDescriptor ObjectDestroyShouldNotBeUsed = new("COIMBRA0019",
                                                                                        "Object.Destroy should be avoided.",
                                                                                        "Use {0}.Destroy() instead of Object.Destroy({0}).",